Add files via upload
this is the code for my workout application for androids. this is my work till now. i am still working on this application.
it includes a welcome screen where you can see the intro and select your plans as a beginner or intermediate. then after selection this shows you a home screen where your working plan is  shown to you and you can start without wasting any time.